// src/app/shared/Models/Geografy.ts
export class PersonGeografy {

    id?: number;
    idPerson?:number;
    idCountry?: number;
    idFederalState?: number;
    idProvince?: number;
    idMunicipality?: number;
    city?: string;
    urbanizationAddress?: string;
    avenueStreet?: string;
    postalCodeAddress?: string;
    numberLocalPhoneCodeAddress?: string;
    numberLocalPhoneAddress?: string;
    numberCelPhoneCodeAddress?: string;
    numberCelPhoneAddress?: string;
    booldelete?: string;
    startIdUser?: string;
    updateIdUser?: string;
    
}